+++ changelog.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Skills and consolidated AGENTS.md
 - Bump org.xerial:sqlite-jdbc from 3.50.3.0 to 3.53.0.0
+- Tons of more code coverage
 
 ### Fixed